Starting TP todolist
