+++ OC-Nodejs/Samples/import.js
@@ -1,0 +1,4 @@
+var monexport = require('./export');
+
+monexport.direBonjour();
+monexport.direByeBye();
