--- OC-Nodejs/app.js
@@ -1,4 +1,0 @@
-var monmodule = require('./monmodule');
-
-monmodule.direBonjour();
-monmodule.direByeBye();
