--- OC-Nodejs/views/chambre.ejs
@@ -1,3 +1,0 @@
-<h1>Vous êtes dans la chambre</h1>
-
-<p>Vous êtes à l'étage n° <%= etage %> </p>
